+++ include/log.h
@@ -20,7 +20,7 @@ void parseLog(std::vector<File>& records, SQLiteHelper& sqliteHelper, std::istre
 
 class LogRecord {
 public:
-  unsigned long long cur_lsn, prev_lsn, undo_lsn;
+  uint64_t cur_lsn, prev_lsn, undo_lsn;
   unsigned int client_id, record_type, flags, redo_op, undo_op, redo_offset, redo_length, undo_offset, undo_length;
   unsigned int target_attr, lcns_to_follow, record_offset, attribute_offset, mft_cluster_index, target_vcn, target_lcn;
   unsigned int client_data_length;
@@ -36,7 +36,7 @@ std::ostream& operator<<(std::ostream& out, const LogRecord& rec);
 class LogData {
 public:
   int64_t Record, Parent, PreviousParent;
-  unsigned long long Lsn;
+  uint64_t Lsn;
   std::string Name, PreviousName, Timestamp;
   std::vector<int> RedoOps, UndoOps;
