+++ include/usn.h
@@ -30,7 +30,7 @@ public:
   void insert(sqlite3_stmt* stmt, const std::vector<File>& records);
   void insertEvent(unsigned int type, sqlite3_stmt* stmt);
 
-  unsigned long long Reference, Record, ParentReference, Parent, PreviousParent, Usn;
+  uint64_t Reference, Record, ParentReference, Parent, PreviousParent, Usn;
   unsigned int Reason;
   bool IsEmbedded;
   std::string Name, PreviousName, Timestamp;
